requirements.txt
# Core AI and ML
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.4.0
langchain==0.1.10
langchain-openai==0.0.8
langchain-anthropic==0.1.4
langchain-google-genai==0.0.11

# Evaluation and Monitoring
opik==1.10.6
opik[fastapi]==1.10.6  # FastAPI integration for automatic tracing
langfuse==2.21.0
wandb==0.16.4

# FastAPI and Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1

# Database and Storage
sqlalchemy==2.0.27
alembic==1.13.1
redis==5.0.1
psycopg2-binary==2.9.9

# Background Tasks and Scheduling
celery==5.3.6
apscheduler==3.10.4

# HTTP and API clients
httpx==0.27.0
requests==2.31.0
aiohttp==3.9.3

# Data Processing
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
scipy==1.12.0

# Behavioral Science and Analytics
statsmodels==0.14.1
plotly==5.18.0
seaborn==0.13.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# Environment and Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-mock==3.12.0
httpx==0.27.0

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Integrations
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.119.0
stripe==8.2.0
twilio==9.0.4

# Monitoring and Logging
structlog==24.1.0
sentry-sdk[fastapi]==1.40.6
prometheus-client==0.20.0